requirements.txt
numpy==2.1.1
Pillow==9.4.0
Pillow==10.4.0
streamlit==1.36.0
tensorflow==2.17.0
tensorflow_intel==2.17.0
